servers/giphy/requirements.txt
# mcp/requirements.txt
fastmcp
pydantic
fastapi
uvicorn[standard]
aiohttp
python-dotenv
typing-extensions